fix(serialization): serialize constraint attributes
