+++ keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/constraints/CanaryConstraintAttributes.kt
@@ -1,0 +1,7 @@
+package com.netflix.spinnaker.keel.api.constraints
+
+data class CanaryConstraintAttributes(
+  val executions: Set<RegionalExecutionId> = emptySet(),
+  val startAttempt: Int = 0,
+  val status: Set<CanaryStatus> = emptySet()
+) : ConstraintStateAttributes("canary")
