+++ keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/constraints/CanaryStatus.kt
@@ -1,0 +1,9 @@
+package com.netflix.spinnaker.keel.api.constraints
+
+data class CanaryStatus(
+  val executionId: String,
+  val region: String,
+  val executionStatus: String,
+  val scores: List<Double> = emptyList(),
+  val scoreMessage: String?
+)
