+++ keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/constraints/PipelineConstraintStateAttributes.kt
@@ -1,0 +1,10 @@
+package com.netflix.spinnaker.keel.api.constraints
+
+import java.time.Instant
+
+data class PipelineConstraintStateAttributes(
+  val executionId: String? = null,
+  val attempt: Int,
+  val latestAttempt: Instant,
+  val lastExecutionStatus: String? = null
+) : ConstraintStateAttributes("pipeline")
