+++ keel-api/src/main/kotlin/com/netflix/spinnaker/keel/api/constraints/RegionalExecutionId.kt
@@ -1,0 +1,6 @@
+package com.netflix.spinnaker.keel.api.constraints
+
+data class RegionalExecutionId(
+  val region: String,
+  val executionId: String
+)
